Game interface updated with team creation and game initiation. Shooter step updated. Teams sorted according to pace.

+++ Seminars/OOP_GAME_PROJECT/src/org/example/units/Magic.java
@@ -5,7 +5,7 @@ public abstract class Magic extends BaseHero{
     protected int maxMana;
     protected int accuracy;
 
-    public Magic(float hp, String name, int x, int y, int armor, int[] damage, int pace, String class_name, int team, int mana, int accuracy) {
+    public Magic(float hp, String name, int x, int y, float armor, int[] damage, int pace, String class_name, int team, int mana, int accuracy) {
         super(hp, name, x, y, armor, damage, pace, class_name, team);
         this.mana = this.maxMana = mana;
         this.accuracy = accuracy;
